+++ src/main/java/net/geforcemods/securitycraft/blocks/CrystalQuartzBlock.java
@@ -5,13 +5,20 @@ import java.util.Random;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.compat.IOverlayDisplay;
 import net.minecraft.block.BlockQuartz;
+import net.minecraft.block.material.MapColor;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 
 public class CrystalQuartzBlock extends BlockQuartz implements IOverlayDisplay {
+	@Override
+	public MapColor getMapColor(IBlockState state, IBlockAccess world, BlockPos pos) {
+		return MapColor.CYAN;
+	}
+
 	@Override
 	public int quantityDropped(Random random) {
 		return 1;