+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedCobwebBlock.java
@@ -17,7 +17,7 @@ import net.minecraft.world.World;
 
 public class ReinforcedCobwebBlock extends BaseReinforcedBlock {
 	public ReinforcedCobwebBlock() {
-		super(Material.GLASS, 1, Blocks.WEB);
+		super(Material.WEB, 1, Blocks.WEB);
 		setSoundType(SoundType.STONE);
 	}