+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedNetherrackBlock.java
@@ -1,9 +1,12 @@
 package net.geforcemods.securitycraft.blocks.reinforced;
 
 import net.minecraft.block.Block;
+import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 
 public class ReinforcedNetherrackBlock extends BaseReinforcedBlock {
@@ -14,5 +17,10 @@ public class ReinforcedNetherrackBlock extends BaseReinforcedBlock {
 	@Override
 	public boolean isFireSource(World world, BlockPos pos, EnumFacing side) {
 		return side == EnumFacing.UP;
+	}
+
+	@Override
+	public MapColor getMapColor(IBlockState state, IBlockAccess worldIn, BlockPos pos) {
+		return MapColor.NETHERRACK;
 	}
 }